backend/pyproject.toml
[project]
name = "link-agent-backend"
version = "0.1.0"
description = "AI-powered LinkedIn content agent backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "psycopg[binary]>=3.2.0",
    "litellm>=1.50.0",
    "google-genai>=1.0.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.12",
    "pypdf>=5.0.0",
    "python-pptx>=1.0.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100
